Now human console takes care of the modification of the support set, so HPEAR is much more cleaner
