+++ scripts/action_recognition_pipeline.py
@@ -1,21 +1,13 @@
-import sys
-from multiprocessing.managers import BaseManager
-from pathlib import Path
 import tensorrt  # Avoid Myelin error (DO NOT REMOVE)
-sys.path.insert(0, Path(__file__).parent.parent.as_posix())
 from configs.action_rec_config import Network, HPE, AR, MAIN
-import os
 import numpy as np
-import time
 import cv2
 import pycuda.autoinit  # Create context on GPU (DO NOT REMOVE)
 
-docker = os.environ.get('AM_I_IN_A_DOCKER_CONTAINER', False)
 
-
-class ISBFSAR(Network.node):
+class HPEAR(Network.node):
     def __init__(self, input_type, cam_width, cam_height, window_size, skeleton_scale, acquisition_time, fps,
-                 consistency_window_length, os_score_thr, detect_focus):
+                 consistency_window_length, os_score_thr):
         super().__init__(**Network.Args.to_dict())
         self.input_type = input_type
         self.cam_width = cam_width
@@ -25,53 +17,43 @@ class ISBFSAR(Network.node):
         self.last_poses = []
         self.last_n_actions = []
         self.skeleton_scale = skeleton_scale
-        self.acquisition_time = acquisition_time
         self.consistency_window_length = consistency_window_length
         self.os_score_thr = os_score_thr
-        self.fps = fps
-        self.detect_focus = detect_focus
-        self.last_time = None
+        # self.fps = fps
         self.edges = None
         self.hpe = None
         self.ar = None
-        self.last_data = None
-        self.commands_queue = None
-        self.last_log = None
 
     def startup(self):
         self.hpe = HPE.model(**HPE.Args.to_dict())
         self.ar = AR.model(**AR.Args.to_dict())
         self.ar.load()
 
-    def get_frame(self, img=None, log=None, cap_fps=True):
-        """
-        get frame, do inference, return all possible info
-        keys: img, bbox, img_preprocessed, human_distance, pose, edges, actions, is_true, requires_focus, focus, face_bbox,
-        fps
-        """
+    def loop(self, data):
+        # Human Console Commands
+        train_data = data["train"] if "train" in data.keys() else None
+        if train_data is not None:
+            self.ar.train(train_data)
+        remove_data = data["remove"] if "remove" in data.keys() else None
+        if remove_data is not None:
+            self.ar.remove(remove_data)
+        debug_data = data["debug"] if "debug" in data.keys() else None
+        if debug_data is not None:
+            self.ar.save_ss_image(self.edges)
+
+        img = data["rgb"]
+        ar_input = {}
         elements = {}
 
-        # If img is not given (not a video), try to get img
-        if img is None:
-            img = self.read("rgb")["rgb"]
-        elements["rgb"] = img
-
-        # Msg
-        if log is not None:
-            self.last_log = log
-        elements["log"] = self.last_log
-
         # Cap fps
-        if self.last_time is not None:
-            while (time.time() - self.last_time) < 1 / self.fps and cap_fps:
-                time.sleep(0.01)
-            self.fps_s.append(1. / (time.time() - self.last_time))
-            fps_s = self.fps_s[-10:]
-            fps = sum(fps_s) / len(fps_s)
-            elements["fps_ar"] = fps
-        self.last_time = time.time()
-
-        ar_input = {}
+        # if self.last_time is not None:
+        #     while (time.time() - self.last_time) < 1 / self.fps and cap_fps:
+        #         time.sleep(0.01)
+        #     self.fps_s.append(1. / (time.time() - self.last_time))
+        #     fps_s = self.fps_s[-10:]
+        #     fps = sum(fps_s) / len(fps_s)
+        #     elements["fps_ar"] = fps
+        # self.last_time = time.time()
 
         # RGB CASE
         hpe_res = self.hpe.estimate(img)
@@ -147,153 +129,7 @@ class ISBFSAR(Network.node):
             # elements["action"] = best_index
         return elements
 
-    def loop(self, data):
-        log = None
-        if data["msg"] is not None:
-            msg = data["msg"].strip()
-            msg = msg.split()
-
-            # select appropriate command
-            if msg[0] == 'close' or msg[0] == 'exit' or msg[0] == 'quit' or msg[0] == 'q':
-                exit()
-
-            elif msg[0] == "add" and len(msg) > 1:
-                log = self.learn_command(msg[1:])
-                data = self.read("rgb")
-
-            elif msg[0] == "remove" and len(msg) > 1:
-                log = self.forget_command(msg[1])
-
-            elif msg[0] == "save":
-                log = self.ar.save()
-
-            elif msg[0] == "load":
-                log = self.ar.load()
-
-            elif msg[0] == "debug":
-                log = self.debug()
-
-            elif msg[0] == "edit_focus":
-                log = self.ar.edit_focus(msg[1], msg[2])
-
-            elif msg[0] == "edit_os":
-                log = self.ar.edit_os(msg[1], msg[2])
-
-            else:
-                log = "Not a valid command!"
-        d = self.get_frame(img=data["rgb"], log=log)
-        return d
-
-    def forget_command(self, flag):
-        if self.ar.remove(flag):
-            return "Action {} removed".format(flag)
-        else:
-            return "Action {} is not in the support set".format(flag)
-
-    def debug(self):
-        ss = self.ar.support_set_data
-
-        labels = self.ar.support_set_labels
-        if len(ss) == 0:
-            return "Support set is empty"
-        if self.input_type in ["hybrid", "rgb"]:
-            ss_rgb = ss["rgb"].detach().cpu().numpy()
-            ss_rgb = ss_rgb.swapaxes(-2, -3).swapaxes(-1, -2)
-            ss_rgb = (ss_rgb - np.array([0.485, 0.456, 0.406])) / np.array([0.229, 0.224, 0.225])
-            ss_rgb = (ss_rgb * 255).astype(np.uint8)
-            way, shot, seq_len, height, width, _ = ss_rgb.shape
-            # Flat image
-            # ss_rgb = ss_rgb.swapaxes(0, 2)
-            # ss_rgb = ss_rgb.reshape(seq_len, shot, way*height, width, 3)
-            sequences = []
-            for w in range(way):
-                for s in range(shot):
-                    support_class = ss_rgb[w][s].swapaxes(0, 1).reshape(height, seq_len * width, 3)
-                    support_class = cv2.putText(support_class, f"{labels[w]}, {s}", (10, 50), cv2.FONT_HERSHEY_SIMPLEX,
-                                                2,
-                                                (255, 255, 255), 3, 2)
-                    sequences.append(support_class)
-            ss_rgb = np.concatenate(sequences, axis=0)
-            cv2.imwrite("SUPPORT_SET.png", ss_rgb)
-        if self.input_type in ["hybrid", "skeleton"]:
-            # ss = np.stack([ss[c]["poses"].detach().cpu().numpy() for c in ss.keys()])
-            classes = []
-            ss_sk = ss["sk"].detach().cpu().numpy()
-            way, shot, _, _ = ss_sk.shape
-            for ss_c in ss_sk:  # FOR EACH CLASS, 5, 16, 90
-                ss_c = ss_c.reshape(ss_c.shape[:-1] + (30, 3))  # 5, 16, 30 , 3
-                size = 250
-                zoom = 2
-                visual = np.zeros((size * ss_c.shape[0], size * ss_c.shape[1]))
-                ss_c = (ss_c + 1) * (size / 2)  # Send each pose from [-1, +1] to [0, size]
-                ss_c *= zoom
-                ss_c = ss_c[..., :2]
-                ss_c[..., 1] += np.arange(ss_c.shape[0])[..., None, None].repeat(ss_c.shape[1], axis=1) * size
-                ss_c[..., 0] += np.arange(ss_c.shape[1])[None, ..., None].repeat(ss_c.shape[0], axis=0) * size
-                ss_c[..., 1] -= size / 2
-                ss_c[..., 0] -= size / 2
-                ss_c = ss_c.reshape(-1, 30, 2).astype(int)
-                for pose in ss_c:
-                    for point in pose:
-                        visual = cv2.circle(visual, point, 1, (255, 0, 0))
-                    for edge in self.edges:
-                        visual = cv2.line(visual, pose[edge[0]], pose[edge[1]], (255, 0, 0))
-                classes.append(visual)
-            visual = np.concatenate(classes, axis=0)
-            for i, label in enumerate(labels):
-                visual = cv2.putText(visual, label, (10, 10 + i * size * shot), cv2.FONT_HERSHEY_SIMPLEX, 1,
-                                     (255, 255, 255), 1, 2)
-            cv2.imwrite("SUPPORT_SET.png", visual)
-        return "Support saved to SUPPORT_SET.png"
-
-    def learn_command(self, flag):
-        action_name = flag[0]
-        try:
-            ss_id = int(flag[1])
-        except Exception:
-            return "Format not valid"
-        requires_focus = len(flag) == 3 and flag[2] == "-focus"
-        now = time.time()
-        while (time.time() - now) < 3:
-            self.write("visualizer", self.get_frame(log="WAIT...", cap_fps=False), False)
-
-        self.write("visualizer", self.get_frame(log="GO!", cap_fps=False), False)
-        data = [[] for _ in range(self.window_size)]
-        i = 0
-        off_time = (self.acquisition_time / self.window_size)
-        while i < self.window_size:
-            start = time.time()
-            res = self.get_frame(log="{:.2f}%".format((i / (self.window_size - 1)) * 100), cap_fps=False)
-            self.write("visualizer", res, False)
-            # Check if the sample is good w.r.t. input type
-            good = self.input_type in ["skeleton", "hybrid"] and "pose" in res.keys() and res["pose"] is not None
-            good = good or self.input_type == "rgb"
-            if good:
-                if self.input_type in ["skeleton", "hybrid"]:
-                    data[i].append(res["pose"].reshape(-1))  # CAREFUL with the reshape
-                if self.input_type in ["rgb", "hybrid"]:
-                    data[i].append(res["img_preprocessed"])
-                i += 1
-            while (time.time() - start) < off_time:  # Busy wait
-                continue
-
-        inp = {"flag": action_name,
-               "data": {},
-               "requires_focus": requires_focus}
-
-        if self.input_type == "rgb":  # Unique case with images in first position
-            inp["data"]["rgb"] = np.stack([x[0] for x in data])
-        if self.input_type in ["skeleton", "hybrid"]:
-            inp["data"]["sk"] = np.stack([x[0] for x in data])
-        if self.input_type == "hybrid":
-            inp["data"]["rgb"] = np.stack([x[1] for x in data])
-        ret = self.ar.train(inp, ss_id)
-        if ret:
-            return "Action " + action_name + " learned successfully!"
-        else:
-            return "Cannot add action"
-
 
 if __name__ == "__main__":
-    m = ISBFSAR(**MAIN.Args.to_dict())
+    m = HPEAR(**MAIN.Args.to_dict())
     m.run()
